+++ internal/lsp/search/reference.go
@@ -35,3 +35,13 @@ func References(doc *ast.APIDoc, uri core.URI, pos core.Position, include bool) 
 
 	return
 }
+
+// Definition 返回所在位置的定义内容
+func Definition(doc *ast.APIDoc, uri core.URI, pos core.Position) core.Location {
+	r := doc.Search(uri, pos, definitionerType)
+	if r == nil {
+		return core.Location{}
+	}
+
+	return r.(ast.Definitioner).Definition().Location
+}